refactor tests for dynamic checks
