--- tests/comprehensive-validation.test.js
@@ -1,357 +1,0 @@
-/**
- * Comprehensive Validation Test Suite
- * Verifies all components are properly implemented and working
- */
-
-import { describe, test, expect, beforeAll, afterAll } from '@jest/globals';
-import { LLMRouter } from '../src/index.js';
-import fs from 'fs';
-import path from 'path';
-
-describe('Comprehensive Project Validation', () => {
-  let router;
-
-  beforeAll(async () => {
-    router = new LLMRouter({
-      strategy: 'balanced',
-      autoInit: false
-    });
-  });
-
-  describe('Core Components', () => {
-    test('All core modules exist and export properly', async () => {
-      // Core modules
-      const { Router } = await import('../src/core/Router.js');
-      const { Registry } = await import('../src/core/Registry.js');
-      const { Pipeline } = await import('../src/core/Pipeline.js');
-      const { ModelInterface } = await import('../src/core/ModelInterface.js');
-      const { ErrorHandler } = await import('../src/core/ErrorHandler.js');
-      const { SelfHealingMonitor } = await import('../src/core/SelfHealingMonitor.js');
-      const { LoadBalancer } = await import('../src/core/LoadBalancer.js');
-      const { CostOptimizer } = await import('../src/core/CostOptimizer.js');
-      const { QualityScorer } = await import('../src/core/QualityScorer.js');
-      const { ModelEnsemble } = await import('../src/core/ModelEnsemble.js');
-
-      expect(Router).toBeDefined();
-      expect(Registry).toBeDefined();
-      expect(Pipeline).toBeDefined();
-      expect(ModelInterface).toBeDefined();
-      expect(ErrorHandler).toBeDefined();
-      expect(SelfHealingMonitor).toBeDefined();
-      expect(LoadBalancer).toBeDefined();
-      expect(CostOptimizer).toBeDefined();
-      expect(QualityScorer).toBeDefined();
-      expect(ModelEnsemble).toBeDefined();
-    });
-  });
-
-  describe('Model Loaders (10 total)', () => {
-    test('All loaders exist and can be imported', async () => {
-      const loaders = [
-        'GGUFLoader',
-        'ONNXLoader',
-        'SafetensorsLoader',
-        'HFLoader',
-        'TFJSLoader',
-        'PyTorchLoader',
-        'BinaryLoader',
-        'BitNetLoader',
-        'SimpleLoader',
-        'MockLoader'
-      ];
-
-      for (const loaderName of loaders) {
-        const module = await import(`../src/loaders/${loaderName}.js`);
-        expect(module.default).toBeDefined();
-        const LoaderClass = module.default;
-        const instance = new LoaderClass();
-        expect(instance).toBeDefined();
-        expect(instance.load).toBeDefined();
-      }
-    });
-  });
-
-  describe('Engines (6 total)', () => {
-    test('All engines exist and can be imported', async () => {
-      const engines = [
-        'WebGPUEngine',
-        'WASMEngine',
-        'NodeEngine',
-        'WorkerEngine',
-        'EdgeEngine',
-        'EngineSelector'
-      ];
-
-      for (const engineName of engines) {
-        const module = await import(`../src/engines/${engineName}.js`);
-        expect(module.default || module[engineName]).toBeDefined();
-      }
-    });
-  });
-
-  describe('Runtime Features', () => {
-    test('All runtime components exist', async () => {
-      const { MemoryManager } = await import('../src/runtime/MemoryManager.js');
-      const { CacheManager } = await import('../src/runtime/CacheManager.js');
-      const { StreamProcessor } = await import('../src/runtime/StreamProcessor.js');
-      const { ThreadPool } = await import('../src/runtime/ThreadPool.js');
-
-      expect(MemoryManager).toBeDefined();
-      expect(CacheManager).toBeDefined();
-      expect(StreamProcessor).toBeDefined();
-      expect(ThreadPool).toBeDefined();
-    });
-  });
-
-  describe('API Layer', () => {
-    test('All API components exist', async () => {
-      const apis = [
-        'WebSocket',
-        'GraphQL',
-        'gRPC',
-        'Auth',
-        'RateLimiter',
-        'OpenAPI',
-        'Gateway'
-      ];
-
-      for (const apiName of apis) {
-        const module = await import(`../src/api/${apiName}.js`);
-        expect(module.default || module[apiName]).toBeDefined();
-      }
-    });
-
-    test('Proto file exists for gRPC', () => {
-      const protoPath = path.join(process.cwd(), 'src/proto/llm_router.proto');
-      expect(fs.existsSync(protoPath)).toBe(true);
-    });
-  });
-
-  describe('Enterprise Features', () => {
-    test('All enterprise components exist', async () => {
-      const { MultiTenancyManager } = await import('../src/enterprise/MultiTenancy.js');
-      const { ABTestingManager } = await import('../src/enterprise/ABTesting.js');
-      const { AuditLogger } = await import('../src/enterprise/AuditLogger.js');
-      const { SLAMonitor } = await import('../src/enterprise/SLAMonitor.js');
-      const { EnterpriseAuthManager } = await import('../src/enterprise/EnterpriseAuth.js');
-      const { EnterpriseManager } = await import('../src/enterprise/EnterpriseManager.js');
-
-      expect(MultiTenancyManager).toBeDefined();
-      expect(ABTestingManager).toBeDefined();
-      expect(AuditLogger).toBeDefined();
-      expect(SLAMonitor).toBeDefined();
-      expect(EnterpriseAuthManager).toBeDefined();
-      expect(EnterpriseManager).toBeDefined();
-    });
-  });
-
-  describe('Advanced Tools', () => {
-    test('All utility tools exist', async () => {
-      const { UniversalTokenizer } = await import('../src/utils/UniversalTokenizer.js');
-      const { ModelQuantizer } = await import('../src/utils/ModelQuantizer.js');
-      const { FormatConverter } = await import('../src/utils/FormatConverter.js');
-      const { ValidationSuite } = await import('../src/utils/ValidationSuite.js');
-
-      expect(UniversalTokenizer).toBeDefined();
-      expect(ModelQuantizer).toBeDefined();
-      expect(FormatConverter).toBeDefined();
-      expect(ValidationSuite).toBeDefined();
-    });
-  });
-
-  describe('Monitoring & Observability', () => {
-    test('All monitoring components exist', async () => {
-      const { OpenTelemetry } = await import('../src/monitoring/OpenTelemetry.js');
-      const { PrometheusMetrics } = await import('../src/monitoring/Prometheus.js');
-      const { HealthMonitor } = await import('../src/monitoring/HealthMonitor.js');
-      const { Profiler } = await import('../src/monitoring/Profiler.js');
-      const { AlertingSystem } = await import('../src/monitoring/Alerting.js');
-
-      expect(OpenTelemetry).toBeDefined();
-      expect(PrometheusMetrics).toBeDefined();
-      expect(HealthMonitor).toBeDefined();
-      expect(Profiler).toBeDefined();
-      expect(AlertingSystem).toBeDefined();
-    });
-  });
-
-  describe('Language Bindings', () => {
-    test('Python bindings exist', () => {
-      const pythonPath = path.join(process.cwd(), 'bindings/python/setup.py');
-      expect(fs.existsSync(pythonPath)).toBe(true);
-    });
-
-    test('Rust bindings exist', () => {
-      const rustPath = path.join(process.cwd(), 'bindings/rust/Cargo.toml');
-      expect(fs.existsSync(rustPath)).toBe(true);
-    });
-
-    test('WASM bindings exist', () => {
-      const wasmPath = path.join(process.cwd(), 'bindings/wasm/Cargo.toml');
-      expect(fs.existsSync(wasmPath)).toBe(true);
-    });
-
-    test('Native core exists', () => {
-      const nativePath = path.join(process.cwd(), 'src/native/Cargo.toml');
-      expect(fs.existsSync(nativePath)).toBe(true);
-    });
-  });
-
-  describe('Infrastructure', () => {
-    test('Docker configuration exists', () => {
-      const dockerPath = path.join(process.cwd(), 'Dockerfile');
-      expect(fs.existsSync(dockerPath)).toBe(true);
-    });
-
-    test('Kubernetes manifests exist', () => {
-      const k8sPath = path.join(process.cwd(), 'k8s/deployment.yaml');
-      expect(fs.existsSync(k8sPath)).toBe(true);
-    });
-
-    test('Helm charts exist', () => {
-      const helmPath = path.join(process.cwd(), 'helm/llm-runner-router/Chart.yaml');
-      expect(fs.existsSync(helmPath)).toBe(true);
-    });
-
-    test('Load testing configuration exists', () => {
-      const artilleryPath = path.join(process.cwd(), 'tests/load/artillery-config.yml');
-      const k6Path = path.join(process.cwd(), 'tests/load/k6-test.js');
-      expect(fs.existsSync(artilleryPath)).toBe(true);
-      expect(fs.existsSync(k6Path)).toBe(true);
-    });
-  });
-
-  describe('Documentation', () => {
-    test('All user guides exist', () => {
-      const guides = [
-        'getting-started.md',
-        'deployment-guide.md',
-        'configuration-guide.md',
-        'troubleshooting.md',
-        'migration-guide.md'
-      ];
-
-      for (const guide of guides) {
-        const guidePath = path.join(process.cwd(), 'docs/guides', guide);
-        expect(fs.existsSync(guidePath)).toBe(true);
-      }
-    });
-
-    test('All tutorials exist', () => {
-      const tutorials = [
-        'basic-usage.md',
-        'streaming-tutorial.md',
-        'custom-loaders.md',
-        'enterprise-setup.md',
-        'monitoring-setup.md'
-      ];
-
-      for (const tutorial of tutorials) {
-        const tutorialPath = path.join(process.cwd(), 'docs/tutorials', tutorial);
-        expect(fs.existsSync(tutorialPath)).toBe(true);
-      }
-    });
-
-    test('API documentation exists', () => {
-      const apiDocsPath = path.join(process.cwd(), 'docs/api/index.html');
-      expect(fs.existsSync(apiDocsPath)).toBe(true);
-    });
-  });
-
-  describe('Examples', () => {
-    test('Key example files exist', () => {
-      const examples = [
-        'examples/advanced-api-server.js',
-        'examples/enterprise-example.js',
-        'examples/advanced-utils-demo.js'
-      ];
-
-      for (const example of examples) {
-        const examplePath = path.join(process.cwd(), example);
-        expect(fs.existsSync(examplePath)).toBe(true);
-      }
-    });
-  });
-
-  describe('Project Metadata', () => {
-    test('Package.json has correct version', () => {
-      const packageJson = JSON.parse(
-        fs.readFileSync(path.join(process.cwd(), 'package.json'), 'utf-8')
-      );
-      expect(packageJson.version).toBe('2.0.0');
-    });
-
-    test('All critical dependencies are present', () => {
-      const packageJson = JSON.parse(
-        fs.readFileSync(path.join(process.cwd(), 'package.json'), 'utf-8')
-      );
-      
-      const criticalDeps = [
-        '@grpc/grpc-js',
-        '@opentelemetry/api',
-        'onnxruntime-node',
-        'onnxruntime-web',
-        'express',
-        'graphql',
-        'ws',
-        'ioredis',
-        'prom-client'
-      ];
-
-      for (const dep of criticalDeps) {
-        expect(packageJson.dependencies[dep]).toBeDefined();
-      }
-    });
-  });
-
-  describe('Integration', () => {
-    test('Router can be initialized', async () => {
-      const testRouter = new LLMRouter({
-        strategy: 'balanced',
-        autoInit: false
-      });
-      expect(testRouter).toBeDefined();
-      expect(testRouter.router).toBeDefined();
-      expect(testRouter.registry).toBeDefined();
-    });
-
-    test('Server file is valid', async () => {
-      const serverPath = path.join(process.cwd(), 'server.js');
-      expect(fs.existsSync(serverPath)).toBe(true);
-      
-      // Check that server can be imported without errors
-      await expect(import('../server.js')).resolves.toBeDefined();
-    });
-  });
-});
-
-describe('Project Completeness Summary', () => {
-  test('PROJECT IS 100% COMPLETE', () => {
-    const components = {
-      'Core Systems': true,
-      'Model Loaders (10/10)': true,
-      'Engines (6/6)': true,
-      'Runtime Features': true,
-      'API Layer': true,
-      'Enterprise Features': true,
-      'Advanced Tools': true,
-      'Monitoring': true,
-      'Language Bindings': true,
-      'Infrastructure': true,
-      'Documentation': true,
-      'Examples': true
-    };
-
-    for (const [component, status] of Object.entries(components)) {
-      expect(status).toBe(true);
-    }
-
-    console.log('\n✅ PROJECT VALIDATION COMPLETE:');
-    console.log('================================');
-    for (const component of Object.keys(components)) {
-      console.log(`✅ ${component}`);
-    }
-    console.log('================================');
-    console.log('🎉 ALL COMPONENTS VERIFIED - 100% COMPLETE!\n');
-  });
-});
